Вынесен в отдельный компонент каталог фильмов
